% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{setURL}
\alias{setURL}
\title{Set the Webhook URL for Sending Messages}
\usage{
setURL(url)
}
\arguments{
\item{url}{A character string specifying the webhook URL.}
}
\value{
Invisible NULL. The function is called for its side effect of setting the URL.
}
\description{
This function sets the webhook URL that will be used by the \code{sendMessage} function.
It must be called before attempting to send any messages.
}
\examples{
\dontrun{
setURL("https://your.webhook.url")
}
}
